Milestone-HMS-V2 Doctor-Dashboard-Management

+++ backend/routes/admin.py
@@ -13,6 +13,9 @@ def is_admin():
 @admin_bp.route('/doctor', methods=['POST'])
 @jwt_required()
 def add_doctor():
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+
     data = request.get_json()
 
     user = User(
@@ -28,7 +31,6 @@ def add_doctor():
         name=data['name'],
         department_id=data['department_id'],
         specialization=data['specialization'],
-        availability=data.get('availability', '')
     )
     db.session.add(doctor)
     db.session.commit()
@@ -37,7 +39,11 @@ def add_doctor():
 
 
 @admin_bp.route('/departments', methods=['GET'])
+@jwt_required()
 def get_departments():
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+    
     departments = Department.query.all()
     return [
         {
@@ -50,6 +56,9 @@ def get_departments():
 @admin_bp.route('/doctor/<int:id>', methods=['GET'])
 @jwt_required()
 def get_doctor(id):
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+    
     doctor = Doctor.query.get(id)
 
     return {
@@ -57,20 +66,21 @@ def get_doctor(id):
         "name": doctor.name,
         "specialization": doctor.specialization,
         "department_id": doctor.department_id,
-        "availability": doctor.availability
     }
 
 
 @admin_bp.route('/doctor/<int:id>', methods=['PUT'])
 @jwt_required()
 def update_doctor(id):
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+    
     doctor = Doctor.query.get(id)
     data = request.get_json()
 
     doctor.name = data.get('name', doctor.name)
     doctor.specialization = data.get('specialization', doctor.specialization)
     doctor.department_id = data.get('department_id', doctor.department_id)
-    doctor.availability = data.get('availability', doctor.availability)
 
     db.session.commit()
 
@@ -145,10 +155,14 @@ def view_appointments():
 
     result = []
     for a in appointments:
+        patient=Patient.query.get(a.patient_id)
+        doctor=Doctor.query.get(a.doctor_id)
         result.append({
             'id': a.id,
             'doctor_id': a.doctor_id,
             'patient_id': a.patient_id,
+            'patient_name':patient.name,
+            'doctor_name':doctor.name,
             'date': a.date,
             'time': a.time,
             'status': a.status
@@ -160,16 +174,28 @@ def view_appointments():
 @admin_bp.route('/delete/user/<int:id>', methods=['DELETE'])
 @jwt_required()
 def delete_user(id):
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+    
     user = User.query.get(id)
 
     if not user:
         return {"message": "User not found"}, 404
 
     doctor = Doctor.query.filter_by(user_id=id).first()
+    patient = Patient.query.filter_by(user_id=id).first()
+
+    if doctor:
+        if Appointment.query.filter_by(doctor_id=doctor.id).first():
+            return {"message": "Cannot delete doctor with appointments"}, 400
+
+    if patient:
+        if Appointment.query.filter_by(patient_id=patient.id).first():
+            return {"message": "Cannot delete patient with appointments"}, 400
+
     if doctor:
         db.session.delete(doctor)
-
-    patient = Patient.query.filter_by(user_id=id).first()
+    
     if patient:
         db.session.delete(patient)
 
@@ -182,6 +208,9 @@ def delete_user(id):
 @admin_bp.route('/blacklist/user/<int:id>', methods=['PUT'])
 @jwt_required()
 def blacklist_user(id):
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+    
     user = User.query.get(id)
 
     if not user:
@@ -196,6 +225,9 @@ def blacklist_user(id):
 @admin_bp.route('/unblacklist/user/<int:id>', methods=['PUT'])
 @jwt_required()
 def unblacklist_user(id):
+    if not is_admin():
+        return {"message": "Unauthorized"}, 403
+    
     user = User.query.get(id)
 
     if not user:
